Moving some work into the base startgg module.
