+++ data/event/build.gradle.kts
@@ -1,12 +1,11 @@
-import com.codingfeline.buildkonfig.compiler.FieldSpec.Type.STRING
-import java.io.FileInputStream
-import java.util.Properties
-
 plugins {
     kotlin("multiplatform")
     id("com.android.library")
     id("com.apollographql.apollo3").version(libs.versions.apollo)
-    id("com.codingfeline.buildkonfig").version("0.13.3")
+}
+
+dependencies {
+    apolloMetadata(project(":data:startgg"))
 }
 
 kotlin {
@@ -20,6 +19,7 @@ kotlin {
                 implementation(project(":data:local-sqldelight"))
                 implementation(project(":data:octanegg"))
                 implementation(project(":data:remote"))
+                implementation(project(":data:startgg"))
                 implementation(libs.apollo.runtime)
                 implementation(libs.kotlinx.coroutines.core)
                 implementation(libs.square.sqldelight.coroutines)
@@ -91,18 +91,6 @@ project.extensions.findByType(org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatfor
 
 apollo {
     service("service") {
-        packageName.set("com.adammcneilly.pocketleague.data.remote.graphql")
-    }
-}
-
-buildkonfig {
-    packageName = "com.adammcneilly.pocketleague.data.event"
-
-    val properties = Properties()
-    properties.load(FileInputStream(project.rootProject.file("local.properties")))
-
-    // default config is required
-    defaultConfigs {
-        buildConfigField(STRING, "START_GG_API_KEY", properties["START_GG_API_KEY"].toString())
+        packageName.set("com.adammcneilly.pocketleague.data.event")
     }
 }
